Server/requirements.txt
agora_token_builder==1.0.0
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.35.22
botocore==1.35.22
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.6.8
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
git-filter-repo==2.45.0
hyperlink==21.0.0
idna==3.9
incremental==24.7.2
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==10.12.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3
